requirements.txt
requests
bs4
streamlit
fastapi
uvicorn
langchain
langchain-community 
pymilvus 
sentence-transformers
python-dotenv
langchain-milvus
openai